+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fixed rendering of custom asset names / symbol in import activities dialog
 
+### Added
+
+- Added support for data gathering by date range in the asset profile details dialog of the admin control panel
+
 ## 2.209.0 - 2025-10-18
 
 ### Added